backend/requirements.txt
celery==5.6.3
Django==6.0.4
django-cors-headers==4.9.0
djangorestframework==3.17.1
psycopg2-binary==2.9.11
redis==7.4.0
gunicorn==23.0.0
